refactor: finish cursor and openrouter provider splits
