+++ internal/providers/cursor/cursor.go
@@ -1,14 +1,6 @@
 package cursor
 
 import (
-	"context"
-	"database/sql"
-	"fmt"
-	"log"
-	"math"
-	"sort"
-	"strconv"
-	"strings"
 	"sync"
 	"time"
 
@@ -16,8 +8,6 @@ import (
 
 	"github.com/janekbaraniewski/openusage/internal/core"
 	"github.com/janekbaraniewski/openusage/internal/providers/providerbase"
-	"github.com/janekbaraniewski/openusage/internal/providers/shared"
-	"github.com/samber/lo"
 )
 
 var cursorAPIBase = "https://api2.cursor.sh"
@@ -171,1755 +161,4 @@ func (p *Provider) now() time.Time {
 		return p.clock.Now()
 	}
 	return time.Now()
-}
-
-func (p *Provider) Fetch(ctx context.Context, acct core.AccountConfig) (core.UsageSnapshot, error) {
-	if strings.TrimSpace(acct.Provider) == "" {
-		acct.Provider = p.ID()
-	}
-	snap := core.UsageSnapshot{
-		ProviderID:  p.ID(),
-		AccountID:   acct.ID,
-		Timestamp:   p.now(),
-		Status:      core.StatusOK,
-		Metrics:     make(map[string]core.Metric),
-		Resets:      make(map[string]time.Time),
-		Raw:         make(map[string]string),
-		DailySeries: make(map[string][]core.TimePoint),
-	}
-	if acct.ExtraData != nil {
-		if email := strings.TrimSpace(acct.ExtraData["email"]); email != "" {
-			snap.Raw["account_email"] = email
-		}
-		if membership := strings.TrimSpace(acct.ExtraData["membership"]); membership != "" {
-			snap.Raw["membership_type"] = membership
-		}
-	}
-
-	acct.NormalizeRuntimePaths()
-	trackingDBPath := acct.Path("tracking_db", "")
-	stateDBPath := acct.Path("state_db", "")
-
-	// If the token was not persisted (json:"-"), try to extract it fresh
-	// from the Cursor state DB so daemon polls can access the API.
-	token := acct.Token
-	if token == "" && stateDBPath != "" {
-		token = extractTokenFromStateDB(stateDBPath)
-	}
-
-	// Run API calls concurrently with local DB reads so heavy local queries
-	// don't consume the context timeout needed by the API.
-	type apiResult struct {
-		snap *core.UsageSnapshot
-		err  error
-	}
-	apiCh := make(chan apiResult, 1)
-	if token != "" {
-		go func() {
-			apiSnap := core.UsageSnapshot{
-				AccountID:   acct.ID,
-				Metrics:     make(map[string]core.Metric),
-				Resets:      make(map[string]time.Time),
-				Raw:         make(map[string]string),
-				DailySeries: make(map[string][]core.TimePoint),
-			}
-			err := p.fetchFromAPI(ctx, token, &apiSnap)
-			apiCh <- apiResult{snap: &apiSnap, err: err}
-		}()
-	} else {
-		apiCh <- apiResult{err: fmt.Errorf("no token")}
-	}
-
-	// Also resolve ExtraData from persisted fields if not present.
-	if acct.ExtraData == nil {
-		acct.ExtraData = make(map[string]string)
-	}
-	if acct.ExtraData["tracking_db"] == "" && trackingDBPath != "" {
-		acct.ExtraData["tracking_db"] = trackingDBPath
-	}
-	if acct.ExtraData["state_db"] == "" && stateDBPath != "" {
-		acct.ExtraData["state_db"] = stateDBPath
-	}
-
-	var hasLocalData bool
-	if trackingDBPath != "" {
-		before := cursorSnapshotDataSignature(&snap)
-		if err := p.readTrackingDB(ctx, trackingDBPath, &snap); err != nil {
-			log.Printf("[cursor] tracking DB error: %v", err)
-			snap.Raw["tracking_db_error"] = err.Error()
-		} else if cursorSnapshotDataSignature(&snap) != before {
-			hasLocalData = true
-		}
-	}
-	if stateDBPath != "" {
-		before := cursorSnapshotDataSignature(&snap)
-		if err := p.readStateDB(ctx, stateDBPath, &snap); err != nil {
-			log.Printf("[cursor] state DB error: %v", err)
-			snap.Raw["state_db_error"] = err.Error()
-		} else if cursorSnapshotDataSignature(&snap) != before {
-			hasLocalData = true
-		}
-	}
-
-	// Collect API results.
-	ar := <-apiCh
-	hasAPIData := false
-	if ar.err == nil && ar.snap != nil {
-		mergeAPIIntoSnapshot(&snap, ar.snap)
-		hasAPIData = true
-	} else if ar.err != nil && token != "" {
-		log.Printf("[cursor] API fetch failed, falling back to local data: %v", ar.err)
-		snap.Raw["api_error"] = ar.err.Error()
-	}
-
-	if !hasAPIData && !hasLocalData {
-		snap.Status = core.StatusError
-		snap.Message = "No Cursor tracking data accessible (no API token and no local DBs)"
-		return snap, nil
-	}
-
-	if !hasAPIData {
-		p.applyCachedModelAggregations(acct.ID, "", "", &snap)
-		p.applyCachedBillingMetrics(acct.ID, &snap)
-		p.buildLocalOnlyMessage(&snap)
-	}
-
-	// Final safety net: ensure credit gauges exist from local data when
-	// API didn't provide them (or API is completely unavailable).
-	p.ensureCreditGauges(acct.ID, &snap)
-
-	return snap, nil
-}
-
-func mergeAPIIntoSnapshot(dst, src *core.UsageSnapshot) {
-	for k, v := range src.Metrics {
-		dst.Metrics[k] = v
-	}
-	for k, v := range src.Resets {
-		dst.Resets[k] = v
-	}
-	for k, v := range src.Raw {
-		dst.Raw[k] = v
-	}
-	for k, v := range src.DailySeries {
-		dst.DailySeries[k] = v
-	}
-	dst.ModelUsage = append(dst.ModelUsage, src.ModelUsage...)
-	if src.Status != "" {
-		dst.Status = src.Status
-	}
-	if src.Message != "" {
-		dst.Message = src.Message
-	}
-}
-
-type cursorSnapshotSignature struct {
-	metrics     int
-	resets      int
-	raw         int
-	dailySeries int
-	modelUsage  int
-}
-
-func cursorSnapshotDataSignature(snap *core.UsageSnapshot) cursorSnapshotSignature {
-	if snap == nil {
-		return cursorSnapshotSignature{}
-	}
-	return cursorSnapshotSignature{
-		metrics:     len(snap.Metrics),
-		resets:      len(snap.Resets),
-		raw:         len(snap.Raw),
-		dailySeries: len(snap.DailySeries),
-		modelUsage:  len(snap.ModelUsage),
-	}
-}
-
-func (p *Provider) buildLocalOnlyMessage(snap *core.UsageSnapshot) {
-	var parts []string
-
-	if m, ok := snap.Metrics["composer_cost"]; ok && m.Used != nil && *m.Used > 0 {
-		parts = append(parts, fmt.Sprintf("$%.2f session cost", *m.Used))
-	}
-	if m, ok := snap.Metrics["total_ai_requests"]; ok && m.Used != nil && *m.Used > 0 {
-		parts = append(parts, fmt.Sprintf("%.0f requests", *m.Used))
-	}
-	if m, ok := snap.Metrics["composer_sessions"]; ok && m.Used != nil && *m.Used > 0 {
-		parts = append(parts, fmt.Sprintf("%.0f sessions", *m.Used))
-	}
-
-	if len(parts) > 0 {
-		snap.Message = strings.Join(parts, " · ") + " (API unavailable)"
-	} else {
-		snap.Message = "Local Cursor IDE usage tracking (API unavailable)"
-	}
-}
-
-func (p *Provider) fetchFromAPI(ctx context.Context, token string, snap *core.UsageSnapshot) error {
-	// All API endpoints are called independently so a single endpoint failure
-	// doesn't lose data from the others.
-	var (
-		hasPeriodUsage                  bool
-		periodUsage                     currentPeriodUsageResp
-		pu                              planUsage
-		su                              spendLimitUsage
-		totalSpendDollars, limitDollars float64
-	)
-	if err := p.callDashboardAPI(ctx, token, "GetCurrentPeriodUsage", &periodUsage); err != nil {
-		log.Printf("[cursor] GetCurrentPeriodUsage failed (continuing with other endpoints): %v", err)
-		snap.Raw["period_usage_error"] = err.Error()
-	} else {
-		hasPeriodUsage = true
-		pu = periodUsage.PlanUsage
-		su = periodUsage.SpendLimitUsage
-		totalSpendDollars = pu.TotalSpend / 100.0
-		includedDollars := pu.IncludedSpend / 100.0
-		limitDollars = pu.Limit / 100.0
-		bonusDollars := pu.BonusSpend / 100.0
-
-		snap.Metrics["plan_spend"] = core.Metric{
-			Used:   &totalSpendDollars,
-			Limit:  &limitDollars,
-			Unit:   "USD",
-			Window: "billing-cycle",
-		}
-		snap.Metrics["plan_included"] = core.Metric{
-			Used:   &includedDollars,
-			Unit:   "USD",
-			Window: "billing-cycle",
-		}
-		snap.Metrics["plan_bonus"] = core.Metric{
-			Used:   &bonusDollars,
-			Unit:   "USD",
-			Window: "billing-cycle",
-		}
-
-		totalPctUsed := pu.TotalPercentUsed
-		totalPctRemaining := 100.0 - totalPctUsed
-		hundredPct := 100.0
-		snap.Metrics["plan_percent_used"] = core.Metric{
-			Used:      &totalPctUsed,
-			Remaining: &totalPctRemaining,
-			Limit:     &hundredPct,
-			Unit:      "%",
-			Window:    "billing-cycle",
-		}
-		autoPctUsed := pu.AutoPercentUsed
-		autoPctRemaining := 100.0 - autoPctUsed
-		snap.Metrics["plan_auto_percent_used"] = core.Metric{
-			Used:      &autoPctUsed,
-			Remaining: &autoPctRemaining,
-			Limit:     &hundredPct,
-			Unit:      "%",
-			Window:    "billing-cycle",
-		}
-		apiPctUsed := pu.APIPercentUsed
-		apiPctRemaining := 100.0 - apiPctUsed
-		snap.Metrics["plan_api_percent_used"] = core.Metric{
-			Used:      &apiPctUsed,
-			Remaining: &apiPctRemaining,
-			Limit:     &hundredPct,
-			Unit:      "%",
-			Window:    "billing-cycle",
-		}
-
-		if su.PooledLimit > 0 {
-			pooledLimitDollars := su.PooledLimit / 100.0
-			pooledUsedDollars := su.PooledUsed / 100.0
-			pooledRemainingDollars := su.PooledRemaining / 100.0
-			individualDollars := su.IndividualUsed / 100.0
-
-			snap.Metrics["spend_limit"] = core.Metric{
-				Limit:     &pooledLimitDollars,
-				Used:      &pooledUsedDollars,
-				Remaining: &pooledRemainingDollars,
-				Unit:      "USD",
-				Window:    "billing-cycle",
-			}
-			snap.Metrics["individual_spend"] = core.Metric{
-				Used:   &individualDollars,
-				Unit:   "USD",
-				Window: "billing-cycle",
-			}
-
-			// Stacked gauge: team_budget shows self vs others within the pooled limit.
-			teamTotalUsedDollars := pooledUsedDollars
-			snap.Metrics["team_budget"] = core.Metric{
-				Limit:  &pooledLimitDollars,
-				Used:   &teamTotalUsedDollars,
-				Unit:   "USD",
-				Window: "billing-cycle",
-			}
-			selfSpend := individualDollars
-			snap.Metrics["team_budget_self"] = core.Metric{
-				Used:   &selfSpend,
-				Unit:   "USD",
-				Window: "billing-cycle",
-			}
-			othersSpend := pooledUsedDollars - individualDollars
-			if othersSpend < 0 {
-				othersSpend = 0
-			}
-			snap.Metrics["team_budget_others"] = core.Metric{
-				Used:   &othersSpend,
-				Unit:   "USD",
-				Window: "billing-cycle",
-			}
-
-			snap.Raw["spend_limit_type"] = su.LimitType
-		}
-
-		snap.Raw["display_message"] = periodUsage.DisplayMessage
-		snap.Raw["display_threshold"] = strconv.FormatFloat(periodUsage.DisplayThreshold, 'f', -1, 64)
-		snap.Raw["billing_cycle_start"] = formatTimestamp(periodUsage.BillingCycleStart)
-		snap.Raw["billing_cycle_end"] = formatTimestamp(periodUsage.BillingCycleEnd)
-
-		cycleStart := shared.FlexParseTime(periodUsage.BillingCycleStart)
-		cycleEnd := shared.FlexParseTime(periodUsage.BillingCycleEnd)
-		if !cycleEnd.IsZero() {
-			snap.Resets["billing_cycle_end"] = cycleEnd
-		}
-		if !cycleStart.IsZero() && !cycleEnd.IsZero() && cycleEnd.After(cycleStart) {
-			totalDuration := cycleEnd.Sub(cycleStart).Seconds()
-			elapsed := time.Since(cycleStart).Seconds()
-			if elapsed < 0 {
-				elapsed = 0
-			}
-			if elapsed > totalDuration {
-				elapsed = totalDuration
-			}
-			cyclePct := (elapsed / totalDuration) * 100
-			remaining := 100.0 - cyclePct
-			hundred := 100.0
-			snap.Metrics["billing_cycle_progress"] = core.Metric{
-				Used:      &cyclePct,
-				Remaining: &remaining,
-				Limit:     &hundred,
-				Unit:      "%",
-				Window:    "billing-cycle",
-			}
-			daysRemaining := cycleEnd.Sub(p.now()).Hours() / 24
-			if daysRemaining < 0 {
-				daysRemaining = 0
-			}
-			snap.Raw["billing_cycle_days_remaining"] = fmt.Sprintf("%.0f", daysRemaining)
-			totalDays := totalDuration / 86400
-			snap.Raw["billing_cycle_total_days"] = fmt.Sprintf("%.0f", totalDays)
-		}
-
-		if su.PooledLimit > 0 && su.PooledRemaining > 0 {
-			spendPctUsed := (su.PooledUsed / su.PooledLimit) * 100
-			if spendPctUsed >= 100 {
-				snap.Status = core.StatusLimited
-			} else if spendPctUsed >= 80 {
-				snap.Status = core.StatusNearLimit
-			}
-		} else if pu.TotalPercentUsed >= 100 {
-			snap.Status = core.StatusLimited
-		} else if pu.TotalPercentUsed >= 80 {
-			snap.Status = core.StatusNearLimit
-		}
-
-		snap.Metrics["plan_total_spend_usd"] = core.Metric{
-			Used:   &totalSpendDollars,
-			Limit:  &limitDollars,
-			Unit:   "USD",
-			Window: "billing-cycle",
-		}
-		if su.PooledLimit > 0 {
-			pooledLimitDollars := su.PooledLimit / 100.0
-			snap.Metrics["plan_limit_usd"] = core.Metric{
-				Limit:  &pooledLimitDollars,
-				Unit:   "USD",
-				Window: "billing-cycle",
-			}
-		} else {
-			snap.Metrics["plan_limit_usd"] = core.Metric{
-				Limit:  &limitDollars,
-				Unit:   "USD",
-				Window: "billing-cycle",
-			}
-		}
-	}
-
-	var planInfo planInfoResp
-	if err := p.callDashboardAPI(ctx, token, "GetPlanInfo", &planInfo); err == nil {
-		snap.Raw["plan_name"] = planInfo.PlanInfo.PlanName
-		snap.Raw["plan_price"] = planInfo.PlanInfo.Price
-		snap.Raw["plan_billing_cycle_end"] = formatTimestamp(planInfo.PlanInfo.BillingCycleEnd)
-		if planInfo.PlanInfo.IncludedAmountCents > 0 {
-			snap.Raw["plan_included_amount_cents"] = strconv.FormatFloat(planInfo.PlanInfo.IncludedAmountCents, 'f', -1, 64)
-			planIncludedAmountUSD := planInfo.PlanInfo.IncludedAmountCents / 100.0
-			snap.Metrics["plan_included_amount"] = core.Metric{
-				Used:   &planIncludedAmountUSD,
-				Unit:   "USD",
-				Window: "billing-cycle",
-			}
-
-			if hasPeriodUsage && limitDollars <= 0 && su.PooledLimit <= 0 {
-				effectiveLimit := planIncludedAmountUSD
-				snap.Metrics["plan_spend"] = core.Metric{
-					Used:   &totalSpendDollars,
-					Limit:  &effectiveLimit,
-					Unit:   "USD",
-					Window: "billing-cycle",
-				}
-			}
-		}
-	}
-
-	effectivePlanLimitUSD := limitDollars
-	if effectivePlanLimitUSD <= 0 && su.PooledLimit > 0 {
-		effectivePlanLimitUSD = su.PooledLimit / 100.0
-	}
-	if effectivePlanLimitUSD <= 0 && planInfo.PlanInfo.IncludedAmountCents > 0 {
-		effectivePlanLimitUSD = planInfo.PlanInfo.IncludedAmountCents / 100.0
-	}
-
-	var aggUsage aggregatedUsageResp
-	aggErr := p.callDashboardAPI(ctx, token, "GetAggregatedUsageEvents", &aggUsage)
-	aggApplied := false
-	if aggErr == nil {
-		aggApplied = applyModelAggregations(snap, aggUsage.Aggregations)
-		if aggApplied {
-			p.storeModelAggregationCache(snap.AccountID, snap.Raw["billing_cycle_start"], snap.Raw["billing_cycle_end"], aggUsage.Aggregations, effectivePlanLimitUSD)
-		}
-		applyAggregationTotals(snap, &aggUsage)
-	}
-	if !aggApplied && p.applyCachedModelAggregations(snap.AccountID, snap.Raw["billing_cycle_start"], snap.Raw["billing_cycle_end"], snap) {
-		if aggErr != nil {
-			log.Printf("[cursor] using cached model aggregation after API error: %v", aggErr)
-		} else {
-			log.Printf("[cursor] using cached model aggregation after empty API aggregation response")
-		}
-	}
-
-	// If GetCurrentPeriodUsage failed but aggregation succeeded, build a
-	// plan_spend gauge from billing_total_cost so credits are visible.
-	if !hasPeriodUsage {
-		p.applyCachedBillingMetrics(snap.AccountID, snap)
-		if _, ok := snap.Metrics["plan_spend"]; !ok {
-			if m, ok := snap.Metrics["billing_total_cost"]; ok && m.Used != nil && *m.Used > 0 {
-				costUSD := *m.Used
-				if effectivePlanLimitUSD > 0 {
-					snap.Metrics["plan_spend"] = core.Metric{
-						Used:   &costUSD,
-						Limit:  core.Float64Ptr(effectivePlanLimitUSD),
-						Unit:   "USD",
-						Window: "billing-cycle",
-					}
-				}
-			}
-		}
-	}
-
-	var hardLimit hardLimitResp
-	if err := p.callDashboardAPI(ctx, token, "GetHardLimit", &hardLimit); err == nil {
-		if hardLimit.NoUsageBasedAllowed {
-			snap.Raw["usage_based_billing"] = "disabled"
-		} else {
-			snap.Raw["usage_based_billing"] = "enabled"
-		}
-	}
-
-	var profile stripeProfileResp
-	if err := p.callRESTAPI(ctx, token, "/auth/full_stripe_profile", &profile); err == nil {
-		snap.Raw["membership_type"] = profile.MembershipType
-		snap.Raw["is_team_member"] = strconv.FormatBool(profile.IsTeamMember)
-		snap.Raw["team_membership"] = profile.TeamMembershipType
-		snap.Raw["individual_membership"] = profile.IndividualMembershipType
-		if profile.IsTeamMember {
-			snap.Raw["team_id"] = fmt.Sprintf("%.0f", profile.TeamID)
-		}
-	}
-
-	var limitPolicy usageLimitPolicyResp
-	if err := p.callDashboardAPI(ctx, token, "GetUsageLimitPolicyStatus", &limitPolicy); err == nil {
-		snap.Raw["can_configure_spend_limit"] = strconv.FormatBool(limitPolicy.CanConfigureSpendLimit)
-		snap.Raw["limit_policy_type"] = limitPolicy.LimitType
-	}
-
-	// Fetch team members if user is on a team.
-	if profile.IsTeamMember && profile.TeamID > 0 {
-		teamIDStr := fmt.Sprintf("%.0f", profile.TeamID)
-		body := []byte(fmt.Sprintf(`{"teamId":"%s"}`, teamIDStr))
-		var teamMembers teamMembersResp
-		if err := p.callDashboardAPIWithBody(ctx, token, "GetTeamMembers", body, &teamMembers); err == nil {
-			var activeCount int
-			var memberNames []string
-			var ownerCount int
-			for _, m := range teamMembers.TeamMembers {
-				if m.IsRemoved {
-					continue
-				}
-				activeCount++
-				memberNames = append(memberNames, m.Name)
-				if strings.Contains(m.Role, "OWNER") {
-					ownerCount++
-				}
-			}
-			teamSize := float64(activeCount)
-			snap.Metrics["team_size"] = core.Metric{Used: &teamSize, Unit: "members", Window: "current"}
-			snap.Raw["team_members"] = strings.Join(memberNames, ", ")
-			snap.Raw["team_size"] = strconv.Itoa(activeCount)
-			if ownerCount > 0 {
-				ownerV := float64(ownerCount)
-				snap.Metrics["team_owners"] = core.Metric{Used: &ownerV, Unit: "owners", Window: "current"}
-			}
-		}
-	}
-
-	planName := snap.Raw["plan_name"]
-	if su.PooledLimit > 0 {
-		pooledLimitDollars := su.PooledLimit / 100.0
-		pooledUsedDollars := su.PooledUsed / 100.0
-		pooledRemainingDollars := su.PooledRemaining / 100.0
-		snap.Message = fmt.Sprintf("%s — $%.0f / $%.0f team spend ($%.0f remaining)",
-			planName, pooledUsedDollars, pooledLimitDollars, pooledRemainingDollars)
-	} else if limitDollars > 0 {
-		snap.Message = fmt.Sprintf("%s — $%.2f / $%.0f plan spend",
-			planName, totalSpendDollars, limitDollars)
-	} else if planName != "" {
-		snap.Message = fmt.Sprintf("%s — %s", planName, periodUsage.DisplayMessage)
-	}
-
-	// Cache billing metrics so credit gauges survive temporary API failures.
-	p.storeBillingMetricsCache(snap.AccountID, snap)
-
-	// If none of the billing/aggregation endpoints yielded useful data,
-	// report an error so the caller knows API data is effectively absent.
-	_, hasPlanSpend := snap.Metrics["plan_spend"]
-	_, hasSpendLimit := snap.Metrics["spend_limit"]
-	_, hasBillingTotal := snap.Metrics["billing_total_cost"]
-	if !hasPlanSpend && !hasSpendLimit && !hasBillingTotal && !hasPeriodUsage && !aggApplied {
-		return fmt.Errorf("all billing API endpoints failed")
-	}
-
-	return nil
-}
-
-func applyModelAggregations(snap *core.UsageSnapshot, aggregations []modelAggregation) bool {
-	if len(aggregations) == 0 {
-		return false
-	}
-	if snap.Metrics == nil {
-		snap.Metrics = make(map[string]core.Metric)
-	}
-	if snap.Raw == nil {
-		snap.Raw = make(map[string]string)
-	}
-
-	var applied bool
-	for _, agg := range aggregations {
-		modelIntent := strings.TrimSpace(agg.ModelIntent)
-		if modelIntent == "" {
-			continue
-		}
-		rec := core.ModelUsageRecord{
-			RawModelID: modelIntent,
-			RawSource:  "api",
-			Window:     "billing-cycle",
-		}
-
-		inputTokens := strings.TrimSpace(agg.InputTokens)
-		outputTokens := strings.TrimSpace(agg.OutputTokens)
-		cacheWriteTokens := strings.TrimSpace(agg.CacheWriteTokens)
-		cacheReadTokens := strings.TrimSpace(agg.CacheReadTokens)
-
-		if agg.TotalCents > 0 {
-			costDollars := agg.TotalCents / 100.0
-			snap.Metrics[fmt.Sprintf("model_%s_cost", modelIntent)] = core.Metric{
-				Used:   &costDollars,
-				Unit:   "USD",
-				Window: "billing-cycle",
-			}
-			rec.CostUSD = core.Float64Ptr(costDollars)
-		}
-		if inputTokens != "" {
-			snap.Raw[fmt.Sprintf("model_%s_input_tokens", modelIntent)] = inputTokens
-		}
-		if outputTokens != "" {
-			snap.Raw[fmt.Sprintf("model_%s_output_tokens", modelIntent)] = outputTokens
-		}
-		if cacheWriteTokens != "" {
-			snap.Raw[fmt.Sprintf("model_%s_cache_write_tokens", modelIntent)] = cacheWriteTokens
-		}
-		if cacheReadTokens != "" {
-			snap.Raw[fmt.Sprintf("model_%s_cache_read_tokens", modelIntent)] = cacheReadTokens
-		}
-		if agg.Tier > 0 {
-			snap.Raw[fmt.Sprintf("model_%s_tier", modelIntent)] = strconv.Itoa(agg.Tier)
-		}
-
-		if parsed, ok := parseModelTokenCount(inputTokens); ok {
-			v := parsed
-			snap.Metrics[fmt.Sprintf("model_%s_input_tokens", modelIntent)] = core.Metric{
-				Used:   &v,
-				Unit:   "tokens",
-				Window: "billing-cycle",
-			}
-			rec.InputTokens = core.Float64Ptr(parsed)
-		}
-		if parsed, ok := parseModelTokenCount(outputTokens); ok {
-			v := parsed
-			snap.Metrics[fmt.Sprintf("model_%s_output_tokens", modelIntent)] = core.Metric{
-				Used:   &v,
-				Unit:   "tokens",
-				Window: "billing-cycle",
-			}
-			rec.OutputTokens = core.Float64Ptr(parsed)
-		}
-		cacheWrite := float64(0)
-		cacheRead := float64(0)
-		hasCacheWrite := false
-		hasCacheRead := false
-		if parsed, ok := parseModelTokenCount(cacheWriteTokens); ok {
-			cacheWrite = parsed
-			hasCacheWrite = true
-			v := parsed
-			snap.Metrics[fmt.Sprintf("model_%s_cache_write_tokens", modelIntent)] = core.Metric{
-				Used:   &v,
-				Unit:   "tokens",
-				Window: "billing-cycle",
-			}
-		}
-		if parsed, ok := parseModelTokenCount(cacheReadTokens); ok {
-			cacheRead = parsed
-			hasCacheRead = true
-			v := parsed
-			snap.Metrics[fmt.Sprintf("model_%s_cache_read_tokens", modelIntent)] = core.Metric{
-				Used:   &v,
-				Unit:   "tokens",
-				Window: "billing-cycle",
-			}
-		}
-		if hasCacheWrite || hasCacheRead {
-			cached := cacheWrite + cacheRead
-			snap.Metrics[fmt.Sprintf("model_%s_cached_tokens", modelIntent)] = core.Metric{
-				Used:   &cached,
-				Unit:   "tokens",
-				Window: "billing-cycle",
-			}
-			rec.CachedTokens = core.Float64Ptr(cached)
-		}
-
-		if agg.TotalCents > 0 || inputTokens != "" || outputTokens != "" || cacheWriteTokens != "" || cacheReadTokens != "" {
-			applied = true
-			snap.AppendModelUsage(rec)
-		}
-	}
-	return applied
-}
-
-func applyAggregationTotals(snap *core.UsageSnapshot, agg *aggregatedUsageResp) {
-	if agg.TotalCostCents > 0 {
-		totalCostUSD := agg.TotalCostCents / 100.0
-		snap.Metrics["billing_total_cost"] = core.Metric{
-			Used:   &totalCostUSD,
-			Unit:   "USD",
-			Window: "billing-cycle",
-		}
-	}
-	if v, ok := parseModelTokenCount(agg.TotalInputTokens); ok {
-		snap.Metrics["billing_input_tokens"] = core.Metric{Used: &v, Unit: "tokens", Window: "billing-cycle"}
-	}
-	if v, ok := parseModelTokenCount(agg.TotalOutputTokens); ok {
-		snap.Metrics["billing_output_tokens"] = core.Metric{Used: &v, Unit: "tokens", Window: "billing-cycle"}
-	}
-	if cwv, cwOK := parseModelTokenCount(agg.TotalCacheWriteTokens); cwOK {
-		if crv, crOK := parseModelTokenCount(agg.TotalCacheReadTokens); crOK {
-			total := cwv + crv
-			snap.Metrics["billing_cached_tokens"] = core.Metric{Used: &total, Unit: "tokens", Window: "billing-cycle"}
-		}
-	}
-}
-
-func parseModelTokenCount(raw string) (float64, bool) {
-	cleaned := strings.TrimSpace(raw)
-	if cleaned == "" {
-		return 0, false
-	}
-	cleaned = strings.ReplaceAll(cleaned, ",", "")
-	cleaned = strings.ReplaceAll(cleaned, "_", "")
-	v, err := strconv.ParseFloat(cleaned, 64)
-	if err != nil {
-		return 0, false
-	}
-	return v, true
-}
-
-func (p *Provider) readTrackingDB(ctx context.Context, dbPath string, snap *core.UsageSnapshot) error {
-	db, err := sql.Open("sqlite3", fmt.Sprintf("file:%s?mode=ro", dbPath))
-	if err != nil {
-		return fmt.Errorf("opening tracking DB: %w", err)
-	}
-	defer db.Close()
-
-	if !cursorTableExists(ctx, db, "ai_code_hashes") {
-		return nil
-	}
-
-	trackingRecords, err := loadTrackingRecords(ctx, db, p.clock)
-	if err != nil {
-		return err
-	}
-	totalRequests := len(trackingRecords)
-	if totalRequests > 0 {
-		total := float64(totalRequests)
-		snap.Metrics["total_ai_requests"] = core.Metric{
-			Used:   &total,
-			Unit:   "requests",
-			Window: "all-time",
-		}
-	}
-
-	today := p.now().Format("2006-01-02")
-	todayCount := 0
-	for _, record := range trackingRecords {
-		if record.OccurredDay == today {
-			todayCount++
-		}
-	}
-	if todayCount > 0 {
-		tc := float64(todayCount)
-		snap.Metrics["requests_today"] = core.Metric{
-			Used:   &tc,
-			Unit:   "requests",
-			Window: "1d",
-		}
-	}
-
-	p.readTrackingSourceBreakdown(trackingRecords, snap, today)
-	p.readTrackingDailyRequests(trackingRecords, snap)
-	p.readTrackingModelBreakdown(trackingRecords, snap, today)
-	p.readTrackingLanguageBreakdown(trackingRecords, snap)
-	p.readScoredCommits(ctx, db, snap)
-	p.readDeletedFiles(ctx, db, snap)
-	p.readTrackedFileContent(ctx, db, snap)
-
-	return nil
-}
-
-func (p *Provider) readScoredCommits(ctx context.Context, db *sql.DB, snap *core.UsageSnapshot) {
-	var totalCommits int
-	if db.QueryRowContext(ctx, `SELECT COUNT(*) FROM scored_commits WHERE linesAdded IS NOT NULL AND linesAdded > 0`).Scan(&totalCommits) != nil || totalCommits == 0 {
-		return
-	}
-
-	rows, err := db.QueryContext(ctx, `
-		SELECT v2AiPercentage, linesAdded, linesDeleted,
-		       tabLinesAdded, tabLinesDeleted,
-		       composerLinesAdded, composerLinesDeleted,
-		       humanLinesAdded, humanLinesDeleted,
-		       blankLinesAdded, blankLinesDeleted
-		FROM scored_commits
-		WHERE linesAdded IS NOT NULL AND linesAdded > 0
-		ORDER BY scoredAt DESC`)
-	if err != nil {
-		return
-	}
-	defer rows.Close()
-
-	var (
-		sumAIPct      float64
-		countWithPct  int
-		totalTabAdd   int
-		totalTabDel   int
-		totalCompAdd  int
-		totalCompDel  int
-		totalHumanAdd int
-		totalHumanDel int
-		totalBlankAdd int
-		totalBlankDel int
-		totalLinesAdd int
-		totalLinesDel int
-	)
-
-	for rows.Next() {
-		var pctStr sql.NullString
-		var linesAdded, linesDeleted sql.NullInt64
-		var tabAdd, tabDel, compAdd, compDel, humanAdd, humanDel sql.NullInt64
-		var blankAdd, blankDel sql.NullInt64
-		if rows.Scan(&pctStr, &linesAdded, &linesDeleted, &tabAdd, &tabDel, &compAdd, &compDel, &humanAdd, &humanDel, &blankAdd, &blankDel) != nil {
-			continue
-		}
-		if pctStr.Valid && pctStr.String != "" {
-			if v, err := strconv.ParseFloat(pctStr.String, 64); err == nil {
-				sumAIPct += v
-				countWithPct++
-			}
-		}
-		if linesAdded.Valid {
-			totalLinesAdd += int(linesAdded.Int64)
-		}
-		if linesDeleted.Valid {
-			totalLinesDel += int(linesDeleted.Int64)
-		}
-		if tabAdd.Valid {
-			totalTabAdd += int(tabAdd.Int64)
-		}
-		if tabDel.Valid {
-			totalTabDel += int(tabDel.Int64)
-		}
-		if compAdd.Valid {
-			totalCompAdd += int(compAdd.Int64)
-		}
-		if compDel.Valid {
-			totalCompDel += int(compDel.Int64)
-		}
-		if humanAdd.Valid {
-			totalHumanAdd += int(humanAdd.Int64)
-		}
-		if humanDel.Valid {
-			totalHumanDel += int(humanDel.Int64)
-		}
-		if blankAdd.Valid {
-			totalBlankAdd += int(blankAdd.Int64)
-		}
-		if blankDel.Valid {
-			totalBlankDel += int(blankDel.Int64)
-		}
-	}
-
-	tc := float64(totalCommits)
-	snap.Metrics["scored_commits"] = core.Metric{Used: &tc, Unit: "commits", Window: "all-time"}
-	snap.Raw["scored_commits_total"] = strconv.Itoa(totalCommits)
-
-	if countWithPct > 0 {
-		avgPct := sumAIPct / float64(countWithPct)
-		avgPct = math.Round(avgPct*10) / 10
-		hundred := 100.0
-		remaining := hundred - avgPct
-		snap.Metrics["ai_code_percentage"] = core.Metric{
-			Used:      &avgPct,
-			Remaining: &remaining,
-			Limit:     &hundred,
-			Unit:      "%",
-			Window:    "all-commits",
-		}
-		snap.Raw["ai_code_pct_avg"] = fmt.Sprintf("%.1f%%", avgPct)
-		snap.Raw["ai_code_pct_sample"] = strconv.Itoa(countWithPct)
-	}
-
-	if totalLinesAdd > 0 || totalLinesDel > 0 {
-		snap.Raw["commit_total_lines_added"] = strconv.Itoa(totalLinesAdd)
-		snap.Raw["commit_total_lines_deleted"] = strconv.Itoa(totalLinesDel)
-	}
-	if totalTabAdd > 0 || totalCompAdd > 0 || totalHumanAdd > 0 {
-		snap.Raw["commit_tab_lines"] = strconv.Itoa(totalTabAdd)
-		snap.Raw["commit_composer_lines"] = strconv.Itoa(totalCompAdd)
-		snap.Raw["commit_human_lines"] = strconv.Itoa(totalHumanAdd)
-	}
-	if totalTabDel > 0 || totalCompDel > 0 || totalHumanDel > 0 {
-		snap.Raw["commit_tab_lines_deleted"] = strconv.Itoa(totalTabDel)
-		snap.Raw["commit_composer_lines_deleted"] = strconv.Itoa(totalCompDel)
-		snap.Raw["commit_human_lines_deleted"] = strconv.Itoa(totalHumanDel)
-	}
-	if totalBlankAdd > 0 || totalBlankDel > 0 {
-		snap.Raw["commit_blank_lines_added"] = strconv.Itoa(totalBlankAdd)
-		snap.Raw["commit_blank_lines_deleted"] = strconv.Itoa(totalBlankDel)
-	}
-}
-
-func (p *Provider) readDeletedFiles(ctx context.Context, db *sql.DB, snap *core.UsageSnapshot) {
-	var count int
-	if db.QueryRowContext(ctx, `SELECT COUNT(*) FROM ai_deleted_files`).Scan(&count) == nil && count > 0 {
-		v := float64(count)
-		snap.Metrics["ai_deleted_files"] = core.Metric{Used: &v, Unit: "files", Window: "all-time"}
-	}
-}
-
-func (p *Provider) readTrackedFileContent(ctx context.Context, db *sql.DB, snap *core.UsageSnapshot) {
-	var count int
-	if db.QueryRowContext(ctx, `SELECT COUNT(*) FROM tracked_file_content`).Scan(&count) == nil && count > 0 {
-		v := float64(count)
-		snap.Metrics["ai_tracked_files"] = core.Metric{Used: &v, Unit: "files", Window: "all-time"}
-	}
-}
-
-func chooseTrackingTimeExpr(ctx context.Context, db *sql.DB) string {
-	columns := cursorTableColumns(ctx, db, "ai_code_hashes")
-	hasCreatedAt := columns["createdat"]
-	hasTimestamp := columns["timestamp"]
-
-	switch {
-	case hasCreatedAt && hasTimestamp:
-		return "COALESCE(createdAt, timestamp)"
-	case hasCreatedAt:
-		return "createdAt"
-	case hasTimestamp:
-		return "timestamp"
-	default:
-		return "0"
-	}
-}
-
-func (p *Provider) readTrackingSourceBreakdown(records []cursorTrackingRecord, snap *core.UsageSnapshot, today string) {
-	clientTotals := map[string]float64{
-		"ide":        0,
-		"cli_agents": 0,
-		"other":      0,
-	}
-	sourceTotals := make(map[string]int)
-	todaySourceTotals := make(map[string]int)
-	var sourceSummary []string
-	for _, record := range records {
-		sourceTotals[record.Source]++
-		if record.OccurredDay == today {
-			todaySourceTotals[record.Source]++
-		}
-	}
-	for source, count := range sourceTotals {
-		value := float64(count)
-		sourceKey := sanitizeCursorMetricName(source)
-		snap.Metrics["source_"+sourceKey+"_requests"] = core.Metric{
-			Used:   &value,
-			Unit:   "requests",
-			Window: "all-time",
-		}
-
-		// Emit interface-level metrics for the Interface breakdown composition.
-		ifaceValue := value
-		snap.Metrics["interface_"+sourceKey] = core.Metric{
-			Used:   &ifaceValue,
-			Unit:   "calls",
-			Window: "all-time",
-		}
-
-		clientKey := cursorClientBucket(source)
-		clientTotals[clientKey] += value
-		sourceSummary = append(sourceSummary, fmt.Sprintf("%s %d", sourceLabel(source), count))
-	}
-
-	if len(sourceSummary) > 0 {
-		snap.Raw["source_usage"] = strings.Join(sourceSummary, " · ")
-	}
-
-	for bucket, value := range clientTotals {
-		if value <= 0 {
-			continue
-		}
-		v := value
-		snap.Metrics["client_"+bucket+"_sessions"] = core.Metric{
-			Used:   &v,
-			Unit:   "sessions",
-			Window: "all-time",
-		}
-	}
-
-	var todaySummary []string
-	for source, count := range todaySourceTotals {
-		if count <= 0 {
-			continue
-		}
-		value := float64(count)
-		sourceKey := sanitizeCursorMetricName(source)
-		snap.Metrics["source_"+sourceKey+"_requests_today"] = core.Metric{
-			Used:   &value,
-			Unit:   "requests",
-			Window: "1d",
-		}
-		todaySummary = append(todaySummary, fmt.Sprintf("%s %d", sourceLabel(source), count))
-	}
-	if len(todaySummary) > 0 {
-		snap.Raw["source_usage_today"] = strings.Join(todaySummary, " · ")
-	}
-}
-
-func (p *Provider) readTrackingDailyRequests(records []cursorTrackingRecord, snap *core.UsageSnapshot) {
-	totalByDay := make(map[string]float64)
-	byClientDay := map[string]map[string]float64{
-		"ide":        make(map[string]float64),
-		"cli_agents": make(map[string]float64),
-		"other":      make(map[string]float64),
-	}
-	bySourceDay := make(map[string]map[string]float64)
-
-	for _, record := range records {
-		day := record.OccurredDay
-		if day == "" {
-			continue
-		}
-		v := 1.0
-		totalByDay[day] += v
-		clientKey := cursorClientBucket(record.Source)
-		byClientDay[clientKey][day] += v
-		sourceKey := sanitizeCursorMetricName(record.Source)
-		if bySourceDay[sourceKey] == nil {
-			bySourceDay[sourceKey] = make(map[string]float64)
-		}
-		bySourceDay[sourceKey][day] += v
-	}
-
-	if len(totalByDay) > 1 {
-		snap.DailySeries["analytics_requests"] = mapToSortedDailyPoints(totalByDay)
-	}
-	for clientKey, pointsByDay := range byClientDay {
-		if len(pointsByDay) < 2 {
-			continue
-		}
-		snap.DailySeries["usage_client_"+clientKey] = mapToSortedDailyPoints(pointsByDay)
-	}
-	for sourceKey, pointsByDay := range bySourceDay {
-		if len(pointsByDay) < 2 {
-			continue
-		}
-		snap.DailySeries["usage_source_"+sourceKey] = mapToSortedDailyPoints(pointsByDay)
-	}
-}
-
-func (p *Provider) readTrackingModelBreakdown(records []cursorTrackingRecord, snap *core.UsageSnapshot, today string) {
-	modelTotals := make(map[string]int)
-	todayModelTotals := make(map[string]int)
-	byModelDay := make(map[string]map[string]float64)
-	var modelSummary []string
-	for _, record := range records {
-		modelTotals[record.Model]++
-		if record.OccurredDay == today {
-			todayModelTotals[record.Model]++
-		}
-		modelKey := sanitizeCursorMetricName(record.Model)
-		if byModelDay[modelKey] == nil {
-			byModelDay[modelKey] = make(map[string]float64)
-		}
-		if record.OccurredDay != "" {
-			byModelDay[modelKey][record.OccurredDay]++
-		}
-	}
-	for model, count := range modelTotals {
-		if count <= 0 {
-			continue
-		}
-
-		value := float64(count)
-		modelKey := sanitizeCursorMetricName(model)
-		snap.Metrics["model_"+modelKey+"_requests"] = core.Metric{
-			Used:   &value,
-			Unit:   "requests",
-			Window: "all-time",
-		}
-		modelSummary = append(modelSummary, fmt.Sprintf("%s %d", sourceLabel(model), count))
-	}
-	if len(modelSummary) > 0 {
-		snap.Raw["model_usage"] = strings.Join(modelSummary, " · ")
-	}
-
-	for model, count := range todayModelTotals {
-		if count <= 0 {
-			continue
-		}
-		modelKey := sanitizeCursorMetricName(model)
-		value := float64(count)
-		snap.Metrics["model_"+modelKey+"_requests_today"] = core.Metric{
-			Used:   &value,
-			Unit:   "requests",
-			Window: "1d",
-		}
-	}
-	for modelKey, pointsByDay := range byModelDay {
-		if len(pointsByDay) < 2 {
-			continue
-		}
-		snap.DailySeries["usage_model_"+modelKey] = mapToSortedDailyPoints(pointsByDay)
-	}
-}
-
-func (p *Provider) readTrackingLanguageBreakdown(records []cursorTrackingRecord, snap *core.UsageSnapshot) {
-	langTotals := make(map[string]int)
-	var langSummary []string
-	for _, record := range records {
-		if strings.TrimSpace(record.FileExt) == "" {
-			continue
-		}
-		langTotals[record.FileExt]++
-	}
-	for ext, count := range langTotals {
-		value := float64(count)
-		langName := extensionToLanguage(ext)
-		langKey := sanitizeCursorMetricName(langName)
-		snap.Metrics["lang_"+langKey] = core.Metric{
-			Used:   &value,
-			Unit:   "requests",
-			Window: "all-time",
-		}
-		langSummary = append(langSummary, fmt.Sprintf("%s %d", langName, count))
-	}
-	if len(langSummary) > 0 {
-		snap.Raw["language_usage"] = strings.Join(langSummary, " · ")
-	}
-}
-
-var extToLang = map[string]string{
-	".ts": "TypeScript", ".tsx": "TypeScript", ".js": "JavaScript", ".jsx": "JavaScript",
-	".py": "Python", ".go": "Go", ".rs": "Rust", ".rb": "Ruby",
-	".java": "Java", ".kt": "Kotlin", ".kts": "Kotlin",
-	".cs": "C#", ".fs": "F#",
-	".cpp": "C++", ".cc": "C++", ".cxx": "C++", ".hpp": "C++",
-	".c": "C", ".h": "C/C++",
-	".swift": "Swift", ".m": "Obj-C",
-	".php": "PHP", ".lua": "Lua", ".r": "R",
-	".scala": "Scala", ".clj": "Clojure", ".ex": "Elixir", ".exs": "Elixir",
-	".hs": "Haskell", ".erl": "Erlang",
-	".html": "HTML", ".htm": "HTML", ".css": "CSS", ".scss": "SCSS", ".less": "LESS",
-	".json": "JSON", ".yaml": "YAML", ".yml": "YAML", ".toml": "TOML", ".xml": "XML",
-	".md": "Markdown", ".mdx": "Markdown",
-	".sql": "SQL", ".graphql": "GraphQL", ".gql": "GraphQL",
-	".sh": "Shell", ".bash": "Shell", ".zsh": "Shell", ".fish": "Shell",
-	".dockerfile": "Docker", ".tf": "Terraform", ".hcl": "HCL",
-	".vue": "Vue", ".svelte": "Svelte", ".astro": "Astro",
-	".dart": "Dart", ".zig": "Zig", ".nim": "Nim", ".v": "V",
-	".proto": "Protobuf", ".wasm": "WASM",
-}
-
-func extensionToLanguage(ext string) string {
-	ext = strings.ToLower(strings.TrimSpace(ext))
-	if !strings.HasPrefix(ext, ".") {
-		ext = "." + ext
-	}
-	if lang, ok := extToLang[ext]; ok {
-		return lang
-	}
-	return strings.TrimPrefix(ext, ".")
-}
-
-func mapToSortedDailyPoints(byDay map[string]float64) []core.TimePoint {
-	if len(byDay) == 0 {
-		return nil
-	}
-	days := lo.Keys(byDay)
-	sort.Strings(days)
-
-	points := make([]core.TimePoint, 0, len(days))
-	for _, day := range days {
-		points = append(points, core.TimePoint{Date: day, Value: byDay[day]})
-	}
-	return points
-}
-
-func cursorClientBucket(source string) string {
-	s := strings.ToLower(strings.TrimSpace(source))
-	switch {
-	case s == "":
-		return "other"
-	case strings.Contains(s, "cloud"), strings.Contains(s, "web"), s == "background-agent", s == "background_agent":
-		return "cloud_agents"
-	case strings.Contains(s, "cli"), strings.Contains(s, "agent"), strings.Contains(s, "terminal"), strings.Contains(s, "cmd"):
-		return "cli_agents"
-	case s == "composer", s == "tab", s == "human", strings.Contains(s, "ide"), strings.Contains(s, "editor"):
-		return "ide"
-	default:
-		return "other"
-	}
-}
-
-func sanitizeCursorMetricName(source string) string {
-	s := strings.ToLower(strings.TrimSpace(source))
-	if s == "" {
-		return "unknown"
-	}
-	var b strings.Builder
-	lastUnderscore := false
-	for _, r := range s {
-		switch {
-		case r >= 'a' && r <= 'z':
-			b.WriteRune(r)
-			lastUnderscore = false
-		case r >= '0' && r <= '9':
-			b.WriteRune(r)
-			lastUnderscore = false
-		default:
-			if !lastUnderscore {
-				b.WriteByte('_')
-				lastUnderscore = true
-			}
-		}
-	}
-	out := strings.Trim(b.String(), "_")
-	if out == "" {
-		return "unknown"
-	}
-	return out
-}
-
-func sourceLabel(source string) string {
-	trimmed := strings.TrimSpace(source)
-	if trimmed == "" {
-		return "unknown"
-	}
-	return trimmed
-}
-
-func (p *Provider) readStateDB(ctx context.Context, dbPath string, snap *core.UsageSnapshot) error {
-	db, err := sql.Open("sqlite3", fmt.Sprintf("file:%s?mode=ro", dbPath))
-	if err != nil {
-		return fmt.Errorf("opening state DB: %w", err)
-	}
-	defer db.Close()
-
-	if err := db.PingContext(ctx); err != nil {
-		return fmt.Errorf("state DB not accessible: %w", err)
-	}
-
-	dailyStatsRecords, err := loadDailyStatsRecords(ctx, db)
-	if err != nil {
-		dailyStatsRecords = nil
-	}
-	composerRecords, err := loadComposerSessionRecords(ctx, db)
-	if err != nil {
-		log.Printf("[cursor] composerData query error: %v", err)
-	}
-	bubbleRecords, err := loadBubbleRecords(ctx, db)
-	if err != nil {
-		log.Printf("[cursor] bubbleId query error: %v", err)
-	}
-
-	p.readDailyStatsToday(dailyStatsRecords, snap)
-	p.readDailyStatsSeries(dailyStatsRecords, snap)
-	p.readComposerSessions(composerRecords, snap)
-	p.readStateMetadata(ctx, db, snap)
-	p.readToolUsage(bubbleRecords, snap)
-
-	return nil
-}
-
-func (p *Provider) readDailyStatsToday(records []cursorDailyStatsRecord, snap *core.UsageSnapshot) {
-	today := p.now().Format("2006-01-02")
-	yesterday := p.now().AddDate(0, 0, -1).Format("2006-01-02")
-	var stats *dailyStats
-	for i := range records {
-		switch records[i].Date {
-		case today:
-			stats = &records[i].Stats
-		case yesterday:
-			if stats == nil {
-				stats = &records[i].Stats
-			}
-		}
-	}
-	if stats == nil {
-		return
-	}
-
-	if stats.TabSuggestedLines > 0 {
-		suggested := float64(stats.TabSuggestedLines)
-		accepted := float64(stats.TabAcceptedLines)
-		snap.Metrics["tab_suggested_lines"] = core.Metric{Used: &suggested, Unit: "lines", Window: "1d"}
-		snap.Metrics["tab_accepted_lines"] = core.Metric{Used: &accepted, Unit: "lines", Window: "1d"}
-	}
-	if stats.ComposerSuggestedLines > 0 {
-		suggested := float64(stats.ComposerSuggestedLines)
-		accepted := float64(stats.ComposerAcceptedLines)
-		snap.Metrics["composer_suggested_lines"] = core.Metric{Used: &suggested, Unit: "lines", Window: "1d"}
-		snap.Metrics["composer_accepted_lines"] = core.Metric{Used: &accepted, Unit: "lines", Window: "1d"}
-	}
-}
-
-func (p *Provider) readComposerSessions(records []cursorComposerSessionRecord, snap *core.UsageSnapshot) {
-	var (
-		totalCostCents     float64
-		totalRequests      int
-		totalSessions      int
-		totalLinesAdded    int
-		totalLinesRemoved  int
-		totalFilesChanged  int
-		totalFilesCreated  int
-		totalFilesRemoved  int
-		agenticSessions    int
-		nonAgenticSessions int
-		totalContextUsed   float64
-		totalContextLimit  float64
-		contextSampleCount int
-		subagentTypes      = make(map[string]int)
-		modelCosts         = make(map[string]float64)
-		modelRequests      = make(map[string]int)
-		modeSessions       = make(map[string]int)
-		forceModes         = make(map[string]int)
-		statusCounts       = make(map[string]int)
-		dailyCost          = make(map[string]float64)
-		dailyRequests      = make(map[string]float64)
-		todayCostCents     float64
-		todayRequests      int
-	)
-
-	now := p.now()
-	todayStart := time.Date(now.Year(), now.Month(), now.Day(), 0, 0, 0, 0, now.Location())
-
-	for _, record := range records {
-		totalSessions++
-		if record.Mode != "" {
-			modeSessions[record.Mode]++
-		}
-		if record.IsAgentic != nil {
-			if *record.IsAgentic {
-				agenticSessions++
-			} else {
-				nonAgenticSessions++
-			}
-		}
-		if record.ForceMode != "" {
-			forceModes[record.ForceMode]++
-		}
-		if record.Status != "" {
-			statusCounts[record.Status]++
-		}
-		totalLinesAdded += record.LinesAdded
-		totalLinesRemoved += record.LinesRemoved
-		if record.FilesChanged > 0 {
-			totalFilesChanged += record.FilesChanged
-		}
-		if record.AddedFiles > 0 {
-			totalFilesCreated += record.AddedFiles
-		}
-		if record.RemovedFiles > 0 {
-			totalFilesRemoved += record.RemovedFiles
-		}
-		if record.ContextTokensUsed > 0 && record.ContextTokenLimit > 0 {
-			totalContextUsed += record.ContextTokensUsed
-			totalContextLimit += record.ContextTokenLimit
-			contextSampleCount++
-		}
-		if record.SubagentType != "" {
-			subagentTypes[record.SubagentType]++
-		}
-
-		var sessionDay string
-		if !record.OccurredAt.IsZero() {
-			sessionDay = record.OccurredAt.In(now.Location()).Format("2006-01-02")
-		}
-
-		for model, mu := range record.Usage {
-			totalCostCents += mu.CostInCents
-			totalRequests += mu.Amount
-			modelCosts[model] += mu.CostInCents
-			modelRequests[model] += mu.Amount
-
-			if sessionDay != "" {
-				dailyCost[sessionDay] += mu.CostInCents
-				dailyRequests[sessionDay] += float64(mu.Amount)
-			}
-			if !record.OccurredAt.IsZero() && record.OccurredAt.After(todayStart) {
-				todayCostCents += mu.CostInCents
-				todayRequests += mu.Amount
-			}
-		}
-	}
-
-	if totalSessions == 0 {
-		return
-	}
-
-	totalCostUSD := totalCostCents / 100.0
-	snap.Metrics["composer_cost"] = core.Metric{
-		Used:   &totalCostUSD,
-		Unit:   "USD",
-		Window: "all-time",
-	}
-
-	if todayCostCents > 0 {
-		todayCostUSD := todayCostCents / 100.0
-		snap.Metrics["today_cost"] = core.Metric{
-			Used:   &todayCostUSD,
-			Unit:   "USD",
-			Window: "1d",
-		}
-	}
-	if todayRequests > 0 {
-		tr := float64(todayRequests)
-		snap.Metrics["today_composer_requests"] = core.Metric{
-			Used:   &tr,
-			Unit:   "requests",
-			Window: "1d",
-		}
-	}
-
-	sessions := float64(totalSessions)
-	snap.Metrics["composer_sessions"] = core.Metric{
-		Used:   &sessions,
-		Unit:   "sessions",
-		Window: "all-time",
-	}
-	reqs := float64(totalRequests)
-	snap.Metrics["composer_requests"] = core.Metric{
-		Used:   &reqs,
-		Unit:   "requests",
-		Window: "all-time",
-	}
-
-	if totalLinesAdded > 0 {
-		la := float64(totalLinesAdded)
-		snap.Metrics["composer_lines_added"] = core.Metric{Used: &la, Unit: "lines", Window: "all-time"}
-	}
-	if totalLinesRemoved > 0 {
-		lr := float64(totalLinesRemoved)
-		snap.Metrics["composer_lines_removed"] = core.Metric{Used: &lr, Unit: "lines", Window: "all-time"}
-	}
-
-	for model, costCents := range modelCosts {
-		costUSD := costCents / 100.0
-		modelKey := sanitizeCursorMetricName(model)
-		snap.Metrics["model_"+modelKey+"_cost"] = core.Metric{
-			Used:   &costUSD,
-			Unit:   "USD",
-			Window: "all-time",
-		}
-		if reqs, ok := modelRequests[model]; ok {
-			r := float64(reqs)
-			if existing, exists := snap.Metrics["model_"+modelKey+"_requests"]; exists && existing.Used != nil {
-				combined := *existing.Used + r
-				snap.Metrics["model_"+modelKey+"_requests"] = core.Metric{
-					Used:   &combined,
-					Unit:   "requests",
-					Window: "all-time",
-				}
-			} else {
-				snap.Metrics["model_"+modelKey+"_requests"] = core.Metric{
-					Used:   &r,
-					Unit:   "requests",
-					Window: "all-time",
-				}
-			}
-		}
-
-		rec := core.ModelUsageRecord{
-			RawModelID: model,
-			RawSource:  "composer",
-			Window:     "all-time",
-			CostUSD:    core.Float64Ptr(costUSD),
-		}
-		if r, ok := modelRequests[model]; ok {
-			rec.Requests = core.Float64Ptr(float64(r))
-		}
-		snap.AppendModelUsage(rec)
-	}
-
-	for mode, count := range modeSessions {
-		v := float64(count)
-		modeKey := sanitizeCursorMetricName(mode)
-		snap.Metrics["mode_"+modeKey+"_sessions"] = core.Metric{
-			Used:   &v,
-			Unit:   "sessions",
-			Window: "all-time",
-		}
-	}
-
-	if totalFilesChanged > 0 {
-		fc := float64(totalFilesChanged)
-		snap.Metrics["composer_files_changed"] = core.Metric{Used: &fc, Unit: "files", Window: "all-time"}
-	}
-	if totalFilesCreated > 0 {
-		v := float64(totalFilesCreated)
-		snap.Metrics["composer_files_created"] = core.Metric{Used: &v, Unit: "files", Window: "all-time"}
-	}
-	if totalFilesRemoved > 0 {
-		v := float64(totalFilesRemoved)
-		snap.Metrics["composer_files_removed"] = core.Metric{Used: &v, Unit: "files", Window: "all-time"}
-	}
-
-	if agenticSessions > 0 {
-		v := float64(agenticSessions)
-		snap.Metrics["agentic_sessions"] = core.Metric{Used: &v, Unit: "sessions", Window: "all-time"}
-	}
-	if nonAgenticSessions > 0 {
-		v := float64(nonAgenticSessions)
-		snap.Metrics["non_agentic_sessions"] = core.Metric{Used: &v, Unit: "sessions", Window: "all-time"}
-	}
-
-	for fm, count := range forceModes {
-		v := float64(count)
-		fmKey := sanitizeCursorMetricName(fm)
-		snap.Metrics["mode_"+fmKey+"_sessions"] = core.Metric{
-			Used:   &v,
-			Unit:   "sessions",
-			Window: "all-time",
-		}
-	}
-
-	if contextSampleCount > 0 {
-		avgPct := (totalContextUsed / totalContextLimit) * 100
-		avgPct = math.Round(avgPct*10) / 10
-		hundred := 100.0
-		remaining := hundred - avgPct
-		snap.Metrics["composer_context_pct"] = core.Metric{
-			Used:      &avgPct,
-			Remaining: &remaining,
-			Limit:     &hundred,
-			Unit:      "%",
-			Window:    "avg",
-		}
-	}
-
-	for saType, count := range subagentTypes {
-		v := float64(count)
-		saKey := sanitizeCursorMetricName(saType)
-		snap.Metrics["subagent_"+saKey+"_sessions"] = core.Metric{
-			Used:   &v,
-			Unit:   "sessions",
-			Window: "all-time",
-		}
-	}
-
-	snap.Raw["composer_total_cost"] = fmt.Sprintf("$%.2f", totalCostUSD)
-	snap.Raw["composer_total_sessions"] = strconv.Itoa(totalSessions)
-	snap.Raw["composer_total_requests"] = strconv.Itoa(totalRequests)
-	if totalLinesAdded > 0 {
-		snap.Raw["composer_lines_added"] = strconv.Itoa(totalLinesAdded)
-		snap.Raw["composer_lines_removed"] = strconv.Itoa(totalLinesRemoved)
-	}
-
-	if len(dailyCost) > 1 {
-		points := make([]core.TimePoint, 0, len(dailyCost))
-		for day, cents := range dailyCost {
-			points = append(points, core.TimePoint{Date: day, Value: cents / 100.0})
-		}
-		sort.Slice(points, func(i, j int) bool { return points[i].Date < points[j].Date })
-		snap.DailySeries["analytics_cost"] = points
-	}
-	if len(dailyRequests) > 1 {
-		points := mapToSortedDailyPoints(dailyRequests)
-		if existing, ok := snap.DailySeries["analytics_requests"]; ok && len(existing) > 0 {
-			snap.DailySeries["analytics_requests"] = mergeDailyPoints(existing, points)
-		} else {
-			snap.DailySeries["composer_requests_daily"] = points
-		}
-	}
-}
-
-func mergeDailyPoints(a, b []core.TimePoint) []core.TimePoint {
-	byDay := make(map[string]float64)
-	for _, p := range a {
-		byDay[p.Date] += p.Value
-	}
-	for _, p := range b {
-		if byDay[p.Date] < p.Value {
-			byDay[p.Date] = p.Value
-		}
-	}
-	return mapToSortedDailyPoints(byDay)
-}
-
-// extractTokenFromStateDB reads the Cursor access token directly from the
-// state.vscdb SQLite database. This is needed because the Token field has
-// json:"-" and is not persisted to the config file, so daemon polls that
-// load accounts from config would otherwise have no API token.
-func extractTokenFromStateDB(dbPath string) string {
-	db, err := sql.Open("sqlite3", fmt.Sprintf("file:%s?mode=ro", dbPath))
-	if err != nil {
-		return ""
-	}
-	defer db.Close()
-
-	var token string
-	if db.QueryRow(`SELECT value FROM ItemTable WHERE key = 'cursorAuth/accessToken'`).Scan(&token) != nil {
-		return ""
-	}
-	return strings.TrimSpace(token)
-}
-
-func (p *Provider) readStateMetadata(ctx context.Context, db *sql.DB, snap *core.UsageSnapshot) {
-	var email string
-	if db.QueryRowContext(ctx,
-		`SELECT value FROM ItemTable WHERE key = 'cursorAuth/cachedEmail'`).Scan(&email) == nil && email != "" {
-		snap.Raw["account_email"] = email
-	}
-
-	var promptCount string
-	if db.QueryRowContext(ctx,
-		`SELECT value FROM ItemTable WHERE key = 'freeBestOfN.promptCount'`).Scan(&promptCount) == nil && promptCount != "" {
-		if v, err := strconv.ParseFloat(promptCount, 64); err == nil && v > 0 {
-			snap.Metrics["total_prompts"] = core.Metric{Used: &v, Unit: "prompts", Window: "all-time"}
-			snap.Raw["total_prompts"] = promptCount
-		}
-	}
-
-	var membership string
-	if db.QueryRowContext(ctx,
-		`SELECT value FROM ItemTable WHERE key = 'cursorAuth/stripeMembershipType'`).Scan(&membership) == nil && membership != "" {
-		if snap.Raw["membership_type"] == "" {
-			snap.Raw["membership_type"] = membership
-		}
-	}
-}
-
-func (p *Provider) readToolUsage(records []cursorBubbleRecord, snap *core.UsageSnapshot) {
-	toolCounts := make(map[string]int)
-	statusCounts := make(map[string]int)
-	var totalCalls int
-
-	for _, record := range records {
-		if strings.TrimSpace(record.ToolName) == "" {
-			continue
-		}
-
-		name := normalizeToolName(record.ToolName)
-		toolCounts[name]++
-		totalCalls++
-
-		if strings.TrimSpace(record.ToolStatus) != "" {
-			statusCounts[record.ToolStatus]++
-		}
-	}
-
-	if totalCalls == 0 {
-		return
-	}
-
-	tc := float64(totalCalls)
-	snap.Metrics["tool_calls_total"] = core.Metric{Used: &tc, Unit: "calls", Window: "all-time"}
-
-	for name, count := range toolCounts {
-		v := float64(count)
-		toolKey := sanitizeCursorMetricName(name)
-		snap.Metrics["tool_"+toolKey] = core.Metric{
-			Used:   &v,
-			Unit:   "calls",
-			Window: "all-time",
-		}
-	}
-
-	if completed, ok := statusCounts["completed"]; ok && completed > 0 {
-		v := float64(completed)
-		snap.Metrics["tool_completed"] = core.Metric{Used: &v, Unit: "calls", Window: "all-time"}
-	}
-	if errored, ok := statusCounts["error"]; ok && errored > 0 {
-		v := float64(errored)
-		snap.Metrics["tool_errored"] = core.Metric{Used: &v, Unit: "calls", Window: "all-time"}
-	}
-	if cancelled, ok := statusCounts["cancelled"]; ok && cancelled > 0 {
-		v := float64(cancelled)
-		snap.Metrics["tool_cancelled"] = core.Metric{Used: &v, Unit: "calls", Window: "all-time"}
-	}
-
-	if totalCalls > 0 {
-		completed := float64(statusCounts["completed"])
-		successPct := (completed / float64(totalCalls)) * 100
-		successPct = math.Round(successPct*10) / 10
-		hundred := 100.0
-		remaining := hundred - successPct
-		snap.Metrics["tool_success_rate"] = core.Metric{
-			Used:      &successPct,
-			Remaining: &remaining,
-			Limit:     &hundred,
-			Unit:      "%",
-			Window:    "all-time",
-		}
-	}
-
-	snap.Raw["tool_calls_total"] = strconv.Itoa(totalCalls)
-	snap.Raw["tool_completed"] = strconv.Itoa(statusCounts["completed"])
-	snap.Raw["tool_errored"] = strconv.Itoa(statusCounts["error"])
-	snap.Raw["tool_cancelled"] = strconv.Itoa(statusCounts["cancelled"])
-}
-
-// normalizeToolName cleans up raw tool names from Cursor bubble data.
-// MCP tools come in formats like:
-//   - "mcp-kubernetes-user-kubernetes-pods_list" (Cursor's internal format)
-//   - Hyphen-prefixed with "user-" for user-installed MCP servers
-//
-// We normalize MCP tools to the canonical "mcp__server__function" format
-// so the telemetry pipeline handles all providers uniformly.
-func normalizeToolName(raw string) string {
-	name := strings.TrimSpace(raw)
-	if name == "" {
-		return "unknown"
-	}
-
-	// Detect MCP tools by prefix.
-	if strings.HasPrefix(name, "mcp-") || strings.HasPrefix(name, "mcp_") {
-		return normalizeCursorMCPName(name)
-	}
-
-	// Strip version suffixes: "read_file_v2" → "read_file"
-	name = strings.TrimSuffix(name, "_v2")
-	name = strings.TrimSuffix(name, "_v3")
-
-	return name
-}
-
-// normalizeCursorMCPName converts Cursor's MCP tool name format to the
-// canonical "mcp__server__function" format used by the telemetry pipeline.
-//
-// Input formats:
-//
-//	"mcp-kubernetes-user-kubernetes-pods_list"  → "mcp__kubernetes__pods_list"
-//	"mcp-notion-workspace-notion-notion-fetch"  → "mcp__notion__fetch"
-//	"mcp_something_else"                        → "mcp__something__else" (fallback)
-func normalizeCursorMCPName(name string) string {
-	// Primary format: "mcp-SERVER-user-SERVER-FUNCTION" (hyphen-separated).
-	if strings.HasPrefix(name, "mcp-") {
-		rest := name[4:] // strip "mcp-"
-		parts := strings.SplitN(rest, "-user-", 2)
-		if len(parts) == 2 {
-			server := parts[0]
-			// After "user-", the server name is repeated then the function follows.
-			// e.g., "kubernetes-pods_list" where "kubernetes" is the repeated server.
-			afterUser := parts[1]
-			// Strip the repeated server prefix if present.
-			serverDash := server + "-"
-			if strings.HasPrefix(afterUser, serverDash) {
-				function := afterUser[len(serverDash):]
-				return "mcp__" + server + "__" + function
-			}
-			// Server not repeated — the whole remainder is server-function.
-			// Try to split on first hyphen: "notion-fetch" → server=notion, function=fetch.
-			if idx := strings.LastIndex(afterUser, "-"); idx > 0 {
-				return "mcp__" + server + "__" + afterUser[idx+1:]
-			}
-			return "mcp__" + server + "__" + afterUser
-		}
-
-		// Simpler format: "mcp-server-function" (no "user" segment).
-		// e.g., "mcp-kubernetes-pods_log"
-		if idx := strings.Index(rest, "-"); idx > 0 {
-			server := rest[:idx]
-			function := rest[idx+1:]
-			return "mcp__" + server + "__" + function
-		}
-		return "mcp__" + rest + "__"
-	}
-
-	// Underscore format: "mcp_server_function" (less common).
-	if strings.HasPrefix(name, "mcp_") {
-		rest := name[4:]
-		if idx := strings.Index(rest, "_"); idx > 0 {
-			server := rest[:idx]
-			function := rest[idx+1:]
-			return "mcp__" + server + "__" + function
-		}
-		return "mcp__" + rest + "__"
-	}
-
-	return name
-}
-
-func (p *Provider) readDailyStatsSeries(records []cursorDailyStatsRecord, snap *core.UsageSnapshot) {
-	for _, record := range records {
-		ds := record.Stats
-		dateStr := record.Date
-		if ds.TabSuggestedLines > 0 || ds.TabAcceptedLines > 0 {
-			snap.DailySeries["tab_suggested"] = append(snap.DailySeries["tab_suggested"],
-				core.TimePoint{Date: dateStr, Value: float64(ds.TabSuggestedLines)})
-			snap.DailySeries["tab_accepted"] = append(snap.DailySeries["tab_accepted"],
-				core.TimePoint{Date: dateStr, Value: float64(ds.TabAcceptedLines)})
-		}
-
-		if ds.ComposerSuggestedLines > 0 || ds.ComposerAcceptedLines > 0 {
-			snap.DailySeries["composer_suggested"] = append(snap.DailySeries["composer_suggested"],
-				core.TimePoint{Date: dateStr, Value: float64(ds.ComposerSuggestedLines)})
-			snap.DailySeries["composer_accepted"] = append(snap.DailySeries["composer_accepted"],
-				core.TimePoint{Date: dateStr, Value: float64(ds.ComposerAcceptedLines)})
-		}
-
-		totalLines := float64(ds.TabSuggestedLines + ds.ComposerSuggestedLines)
-		if totalLines > 0 {
-			snap.DailySeries["total_lines"] = append(snap.DailySeries["total_lines"],
-				core.TimePoint{Date: dateStr, Value: totalLines})
-		}
-	}
-}
-
-func formatTimestamp(s string) string {
-	t := shared.FlexParseTime(s)
-	if t.IsZero() {
-		return s // return as-is if we can't parse
-	}
-	return t.Format("Jan 02, 2006 15:04 MST")
 }
